win32 fixes
Originally committed as revision 84 to svn://svn.ffmpeg.org/ffmpeg/trunk

+++ libavcodec/mpegaudio.h
@@ -19,7 +19,7 @@ typedef struct MpegAudioContext {
     int bitrate_index; /* bit rate */
     int freq_index;
     int frame_size; /* frame size, in bits, without padding */
-    long long nb_samples; /* total number of samples encoded */
+    INT64 nb_samples; /* total number of samples encoded */
     /* padding computation */
     int frame_frac, frame_frac_incr, do_padding;
     short samples_buf[MPA_MAX_CHANNELS][SAMPLES_BUF_SIZE]; /* buffer for filter */
